collect indexing tests symmetrically
