+++ pandas/tests/indexes/datetimes/test_datetime.py
@@ -2,7 +2,7 @@
 import pytest
 
 import numpy as np
-from datetime import date, timedelta, time
+from datetime import date
 
 import dateutil
 import pandas as pd
@@ -17,112 +17,6 @@ randn = np.random.randn
 
 
 class TestDatetimeIndex(object):
-
-    def test_get_loc(self):
-        idx = pd.date_range('2000-01-01', periods=3)
-
-        for method in [None, 'pad', 'backfill', 'nearest']:
-            assert idx.get_loc(idx[1], method) == 1
-            assert idx.get_loc(idx[1].to_pydatetime(), method) == 1
-            assert idx.get_loc(str(idx[1]), method) == 1
-
-            if method is not None:
-                assert idx.get_loc(idx[1], method,
-                                   tolerance=pd.Timedelta('0 days')) == 1
-
-        assert idx.get_loc('2000-01-01', method='nearest') == 0
-        assert idx.get_loc('2000-01-01T12', method='nearest') == 1
-
-        assert idx.get_loc('2000-01-01T12', method='nearest',
-                           tolerance='1 day') == 1
-        assert idx.get_loc('2000-01-01T12', method='nearest',
-                           tolerance=pd.Timedelta('1D')) == 1
-        assert idx.get_loc('2000-01-01T12', method='nearest',
-                           tolerance=np.timedelta64(1, 'D')) == 1
-        assert idx.get_loc('2000-01-01T12', method='nearest',
-                           tolerance=timedelta(1)) == 1
-        with tm.assert_raises_regex(ValueError,
-                                    'unit abbreviation w/o a number'):
-            idx.get_loc('2000-01-01T12', method='nearest', tolerance='foo')
-        with pytest.raises(KeyError):
-            idx.get_loc('2000-01-01T03', method='nearest', tolerance='2 hours')
-        with pytest.raises(
-                ValueError,
-                match='tolerance size must match target index size'):
-            idx.get_loc('2000-01-01', method='nearest',
-                        tolerance=[pd.Timedelta('1day').to_timedelta64(),
-                                   pd.Timedelta('1day').to_timedelta64()])
-
-        assert idx.get_loc('2000', method='nearest') == slice(0, 3)
-        assert idx.get_loc('2000-01', method='nearest') == slice(0, 3)
-
-        assert idx.get_loc('1999', method='nearest') == 0
-        assert idx.get_loc('2001', method='nearest') == 2
-
-        with pytest.raises(KeyError):
-            idx.get_loc('1999', method='pad')
-        with pytest.raises(KeyError):
-            idx.get_loc('2001', method='backfill')
-
-        with pytest.raises(KeyError):
-            idx.get_loc('foobar')
-        with pytest.raises(TypeError):
-            idx.get_loc(slice(2))
-
-        idx = pd.to_datetime(['2000-01-01', '2000-01-04'])
-        assert idx.get_loc('2000-01-02', method='nearest') == 0
-        assert idx.get_loc('2000-01-03', method='nearest') == 1
-        assert idx.get_loc('2000-01', method='nearest') == slice(0, 2)
-
-        # time indexing
-        idx = pd.date_range('2000-01-01', periods=24, freq='H')
-        tm.assert_numpy_array_equal(idx.get_loc(time(12)),
-                                    np.array([12]), check_dtype=False)
-        tm.assert_numpy_array_equal(idx.get_loc(time(12, 30)),
-                                    np.array([]), check_dtype=False)
-        with pytest.raises(NotImplementedError):
-            idx.get_loc(time(12, 30), method='pad')
-
-    def test_get_indexer(self):
-        idx = pd.date_range('2000-01-01', periods=3)
-        exp = np.array([0, 1, 2], dtype=np.intp)
-        tm.assert_numpy_array_equal(idx.get_indexer(idx), exp)
-
-        target = idx[0] + pd.to_timedelta(['-1 hour', '12 hours',
-                                           '1 day 1 hour'])
-        tm.assert_numpy_array_equal(idx.get_indexer(target, 'pad'),
-                                    np.array([-1, 0, 1], dtype=np.intp))
-        tm.assert_numpy_array_equal(idx.get_indexer(target, 'backfill'),
-                                    np.array([0, 1, 2], dtype=np.intp))
-        tm.assert_numpy_array_equal(idx.get_indexer(target, 'nearest'),
-                                    np.array([0, 1, 1], dtype=np.intp))
-        tm.assert_numpy_array_equal(
-            idx.get_indexer(target, 'nearest',
-                            tolerance=pd.Timedelta('1 hour')),
-            np.array([0, -1, 1], dtype=np.intp))
-        tol_raw = [pd.Timedelta('1 hour'),
-                   pd.Timedelta('1 hour'),
-                   pd.Timedelta('1 hour').to_timedelta64(), ]
-        tm.assert_numpy_array_equal(
-            idx.get_indexer(target, 'nearest',
-                            tolerance=[np.timedelta64(x) for x in tol_raw]),
-            np.array([0, -1, 1], dtype=np.intp))
-        tol_bad = [pd.Timedelta('2 hour').to_timedelta64(),
-                   pd.Timedelta('1 hour').to_timedelta64(),
-                   'foo', ]
-        with pytest.raises(
-                ValueError, match='abbreviation w/o a number'):
-            idx.get_indexer(target, 'nearest', tolerance=tol_bad)
-        with pytest.raises(ValueError):
-            idx.get_indexer(idx[[0]], method='nearest', tolerance='foo')
-
-    def test_reasonable_keyerror(self):
-        # GH #1062
-        index = DatetimeIndex(['1/3/2000'])
-        try:
-            index.get_loc('1/1/2000')
-        except KeyError as e:
-            assert '2000' in str(e)
 
     def test_roundtrip_pickle_with_tz(self):
