+++ pandas/tests/indexes/timedeltas/test_indexing.py
@@ -8,6 +8,149 @@ import pandas.util.testing as tm
 from pandas import TimedeltaIndex, timedelta_range, compat, Index, Timedelta
 
 
+class TestGetItem(object):
+    def test_getitem(self):
+        idx1 = timedelta_range('1 day', '31 day', freq='D', name='idx')
+
+        for idx in [idx1]:
+            result = idx[0]
+            assert result == Timedelta('1 day')
+
+            result = idx[0:5]
+            expected = timedelta_range('1 day', '5 day', freq='D',
+                                       name='idx')
+            tm.assert_index_equal(result, expected)
+            assert result.freq == expected.freq
+
+            result = idx[0:10:2]
+            expected = timedelta_range('1 day', '9 day', freq='2D',
+                                       name='idx')
+            tm.assert_index_equal(result, expected)
+            assert result.freq == expected.freq
+
+            result = idx[-20:-5:3]
+            expected = timedelta_range('12 day', '24 day', freq='3D',
+                                       name='idx')
+            tm.assert_index_equal(result, expected)
+            assert result.freq == expected.freq
+
+            result = idx[4::-1]
+            expected = TimedeltaIndex(['5 day', '4 day', '3 day',
+                                       '2 day', '1 day'],
+                                      freq='-1D', name='idx')
+            tm.assert_index_equal(result, expected)
+            assert result.freq == expected.freq
+
+
+class TestWhere(object):
+    # placeholder for symmetry with DatetimeIndex and PeriodIndex tests
+    pass
+
+
+class TestTake(object):
+    def test_take(self):
+        # GH 10295
+        idx1 = timedelta_range('1 day', '31 day', freq='D', name='idx')
+
+        for idx in [idx1]:
+            result = idx.take([0])
+            assert result == Timedelta('1 day')
+
+            result = idx.take([-1])
+            assert result == Timedelta('31 day')
+
+            result = idx.take([0, 1, 2])
+            expected = timedelta_range('1 day', '3 day', freq='D',
+                                       name='idx')
+            tm.assert_index_equal(result, expected)
+            assert result.freq == expected.freq
+
+            result = idx.take([0, 2, 4])
+            expected = timedelta_range('1 day', '5 day', freq='2D',
+                                       name='idx')
+            tm.assert_index_equal(result, expected)
+            assert result.freq == expected.freq
+
+            result = idx.take([7, 4, 1])
+            expected = timedelta_range('8 day', '2 day', freq='-3D',
+                                       name='idx')
+            tm.assert_index_equal(result, expected)
+            assert result.freq == expected.freq
+
+            result = idx.take([3, 2, 5])
+            expected = TimedeltaIndex(['4 day', '3 day', '6 day'], name='idx')
+            tm.assert_index_equal(result, expected)
+            assert result.freq is None
+
+            result = idx.take([-3, 2, 5])
+            expected = TimedeltaIndex(['29 day', '3 day', '6 day'], name='idx')
+            tm.assert_index_equal(result, expected)
+            assert result.freq is None
+
+    def test_take_invalid_kwargs(self):
+        idx = timedelta_range('1 day', '31 day', freq='D', name='idx')
+        indices = [1, 6, 5, 9, 10, 13, 15, 3]
+
+        msg = r"take\(\) got an unexpected keyword argument 'foo'"
+        tm.assert_raises_regex(TypeError, msg, idx.take,
+                               indices, foo=2)
+
+        msg = "the 'out' parameter is not supported"
+        tm.assert_raises_regex(ValueError, msg, idx.take,
+                               indices, out=indices)
+
+        msg = "the 'mode' parameter is not supported"
+        tm.assert_raises_regex(ValueError, msg, idx.take,
+                               indices, mode='clip')
+
+    # TODO: This method came from test_timedelta; de-dup with version above
+    def test_take2(self):
+        tds = ['1day 02:00:00', '1 day 04:00:00', '1 day 10:00:00']
+        idx = TimedeltaIndex(start='1d', end='2d', freq='H', name='idx')
+        expected = TimedeltaIndex(tds, freq=None, name='idx')
+
+        taken1 = idx.take([2, 4, 10])
+        taken2 = idx[[2, 4, 10]]
+
+        for taken in [taken1, taken2]:
+            tm.assert_index_equal(taken, expected)
+            assert isinstance(taken, TimedeltaIndex)
+            assert taken.freq is None
+            assert taken.name == expected.name
+
+    def test_take_fill_value(self):
+        # GH 12631
+        idx = TimedeltaIndex(['1 days', '2 days', '3 days'],
+                             name='xxx')
+        result = idx.take(np.array([1, 0, -1]))
+        expected = TimedeltaIndex(['2 days', '1 days', '3 days'],
+                                  name='xxx')
+        tm.assert_index_equal(result, expected)
+
+        # fill_value
+        result = idx.take(np.array([1, 0, -1]), fill_value=True)
+        expected = TimedeltaIndex(['2 days', '1 days', 'NaT'],
+                                  name='xxx')
+        tm.assert_index_equal(result, expected)
+
+        # allow_fill=False
+        result = idx.take(np.array([1, 0, -1]), allow_fill=False,
+                          fill_value=True)
+        expected = TimedeltaIndex(['2 days', '1 days', '3 days'],
+                                  name='xxx')
+        tm.assert_index_equal(result, expected)
+
+        msg = ('When allow_fill=True and fill_value is not None, '
+               'all indices must be >= -1')
+        with tm.assert_raises_regex(ValueError, msg):
+            idx.take(np.array([1, 0, -2]), fill_value=True)
+        with tm.assert_raises_regex(ValueError, msg):
+            idx.take(np.array([1, 0, -5]), fill_value=True)
+
+        with pytest.raises(IndexError):
+            idx.take(np.array([1, -5]))
+
+
 class TestTimedeltaIndex(object):
 
     def test_insert(self):
@@ -117,140 +260,6 @@ class TestTimedeltaIndex(object):
             tm.assert_index_equal(result, expected)
             assert result.name == expected.name
             assert result.freq == expected.freq
-
-    def test_getitem(self):
-        idx1 = timedelta_range('1 day', '31 day', freq='D', name='idx')
-
-        for idx in [idx1]:
-            result = idx[0]
-            assert result == Timedelta('1 day')
-
-            result = idx[0:5]
-            expected = timedelta_range('1 day', '5 day', freq='D',
-                                       name='idx')
-            tm.assert_index_equal(result, expected)
-            assert result.freq == expected.freq
-
-            result = idx[0:10:2]
-            expected = timedelta_range('1 day', '9 day', freq='2D',
-                                       name='idx')
-            tm.assert_index_equal(result, expected)
-            assert result.freq == expected.freq
-
-            result = idx[-20:-5:3]
-            expected = timedelta_range('12 day', '24 day', freq='3D',
-                                       name='idx')
-            tm.assert_index_equal(result, expected)
-            assert result.freq == expected.freq
-
-            result = idx[4::-1]
-            expected = TimedeltaIndex(['5 day', '4 day', '3 day',
-                                       '2 day', '1 day'],
-                                      freq='-1D', name='idx')
-            tm.assert_index_equal(result, expected)
-            assert result.freq == expected.freq
-
-    def test_take(self):
-        # GH 10295
-        idx1 = timedelta_range('1 day', '31 day', freq='D', name='idx')
-
-        for idx in [idx1]:
-            result = idx.take([0])
-            assert result == Timedelta('1 day')
-
-            result = idx.take([-1])
-            assert result == Timedelta('31 day')
-
-            result = idx.take([0, 1, 2])
-            expected = timedelta_range('1 day', '3 day', freq='D',
-                                       name='idx')
-            tm.assert_index_equal(result, expected)
-            assert result.freq == expected.freq
-
-            result = idx.take([0, 2, 4])
-            expected = timedelta_range('1 day', '5 day', freq='2D',
-                                       name='idx')
-            tm.assert_index_equal(result, expected)
-            assert result.freq == expected.freq
-
-            result = idx.take([7, 4, 1])
-            expected = timedelta_range('8 day', '2 day', freq='-3D',
-                                       name='idx')
-            tm.assert_index_equal(result, expected)
-            assert result.freq == expected.freq
-
-            result = idx.take([3, 2, 5])
-            expected = TimedeltaIndex(['4 day', '3 day', '6 day'], name='idx')
-            tm.assert_index_equal(result, expected)
-            assert result.freq is None
-
-            result = idx.take([-3, 2, 5])
-            expected = TimedeltaIndex(['29 day', '3 day', '6 day'], name='idx')
-            tm.assert_index_equal(result, expected)
-            assert result.freq is None
-
-    def test_take_invalid_kwargs(self):
-        idx = timedelta_range('1 day', '31 day', freq='D', name='idx')
-        indices = [1, 6, 5, 9, 10, 13, 15, 3]
-
-        msg = r"take\(\) got an unexpected keyword argument 'foo'"
-        tm.assert_raises_regex(TypeError, msg, idx.take,
-                               indices, foo=2)
-
-        msg = "the 'out' parameter is not supported"
-        tm.assert_raises_regex(ValueError, msg, idx.take,
-                               indices, out=indices)
-
-        msg = "the 'mode' parameter is not supported"
-        tm.assert_raises_regex(ValueError, msg, idx.take,
-                               indices, mode='clip')
-
-    # TODO: This method came from test_timedelta; de-dup with version above
-    def test_take2(self):
-        tds = ['1day 02:00:00', '1 day 04:00:00', '1 day 10:00:00']
-        idx = TimedeltaIndex(start='1d', end='2d', freq='H', name='idx')
-        expected = TimedeltaIndex(tds, freq=None, name='idx')
-
-        taken1 = idx.take([2, 4, 10])
-        taken2 = idx[[2, 4, 10]]
-
-        for taken in [taken1, taken2]:
-            tm.assert_index_equal(taken, expected)
-            assert isinstance(taken, TimedeltaIndex)
-            assert taken.freq is None
-            assert taken.name == expected.name
-
-    def test_take_fill_value(self):
-        # GH 12631
-        idx = TimedeltaIndex(['1 days', '2 days', '3 days'],
-                             name='xxx')
-        result = idx.take(np.array([1, 0, -1]))
-        expected = TimedeltaIndex(['2 days', '1 days', '3 days'],
-                                  name='xxx')
-        tm.assert_index_equal(result, expected)
-
-        # fill_value
-        result = idx.take(np.array([1, 0, -1]), fill_value=True)
-        expected = TimedeltaIndex(['2 days', '1 days', 'NaT'],
-                                  name='xxx')
-        tm.assert_index_equal(result, expected)
-
-        # allow_fill=False
-        result = idx.take(np.array([1, 0, -1]), allow_fill=False,
-                          fill_value=True)
-        expected = TimedeltaIndex(['2 days', '1 days', '3 days'],
-                                  name='xxx')
-        tm.assert_index_equal(result, expected)
-
-        msg = ('When allow_fill=True and fill_value is not None, '
-               'all indices must be >= -1')
-        with tm.assert_raises_regex(ValueError, msg):
-            idx.take(np.array([1, 0, -2]), fill_value=True)
-        with tm.assert_raises_regex(ValueError, msg):
-            idx.take(np.array([1, 0, -5]), fill_value=True)
-
-        with pytest.raises(IndexError):
-            idx.take(np.array([1, -5]))
 
     def test_get_loc(self):
         idx = pd.to_timedelta(['0 days', '1 days', '2 days'])
